+++ src/models/InventoryIntegration.ts
@@ -1,0 +1,5 @@
+
+export interface InventoryIntegration {
+    DeliveryConfigID?: string
+    NotificationPointOnly?: boolean
+}