+++ src/models/ProductCollectionProduct.ts
@@ -1,0 +1,29 @@
+import { Inventory } from './Inventory';
+
+export interface ProductCollectionProduct<TProductCollectionProductXp = any> {
+    ListOrder?: number
+    DateAdded?: string
+    OwnerID?: string
+    DefaultPriceScheduleID?: string
+    AutoForward?: boolean
+    ID?: string
+    ParentID?: string
+    IsParent?: boolean
+    readonly IsBundle?: boolean
+    Name?: string
+    Description?: string
+    QuantityMultiplier?: number
+    ShipWeight?: number
+    ShipHeight?: number
+    ShipWidth?: number
+    ShipLength?: number
+    Active?: boolean
+    readonly SpecCount?: number
+    readonly VariantCount?: number
+    ShipFromAddressID?: string
+    Inventory?: Inventory
+    DefaultSupplierID?: string
+    AllSuppliersCanSell?: boolean
+    Returnable?: boolean
+    xp?: TProductCollectionProductXp
+}