+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the ordercloud-javascript-sdk will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [9.0.0] - 2024-12-05
+- This version includes a breaking change that updates the following return types:
+- `ProductCollectionBuyerProduct` (updated from `BuyerProduct`) is the new type returned from `Me.ListProductCollectionEntries` 
+- `ProductCollectionProduct` (updated from `Product`) is the new type returned from `ProductCollections.ListEntries`
+
 # [8.0.2] - 2024-12-03
 - Bug fix for "process is undefined" accessor error when not using NextJS